backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
alembic==1.13.1
psycopg2-binary
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
sendgrid==6.11.0
aiofiles>=23.0.0
ruff>=0.4.4
pytest>=8.2.0
pytest-asyncio>=0.23.6
